refactor(170): reorganize domain puzzle validation tests
