+++ app/src/test/java/org/cescfe/numpairs/domain/puzzle/validation/SumProductPairingTest.kt
@@ -1,4 +1,4 @@
-package org.cescfe.numpairs.domain.puzzle
+package org.cescfe.numpairs.domain.puzzle.validation
 
 import org.cescfe.numpairs.domain.puzzle.model.Operator
 import org.cescfe.numpairs.domain.puzzle.model.Strip
@@ -7,8 +7,6 @@ import org.cescfe.numpairs.domain.puzzle.model.TileResolutionState
 import org.cescfe.numpairs.domain.puzzle.support.TileAssignment
 import org.cescfe.numpairs.domain.puzzle.support.defaultKnownStripValues
 import org.cescfe.numpairs.domain.puzzle.support.knownPuzzleWithAssignments
-import org.cescfe.numpairs.domain.puzzle.validation.hasMismatchedSumProductPairings
-import org.cescfe.numpairs.domain.puzzle.validation.mismatchedSumProductPairingTileIndexes
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
